Update [04/03/2020]
- After compiling with -Wall the small warnings appeared in 
mod_vars.F90, mod_seed.F90, mod_print.F90 and mod_calendar.F90 are 
fixed.

- mod_vertvel.F90: use generic variables ix, ixm, jy and kz to avoid 
conflicts.

+++ src/mod_calendar.F90
@@ -311,7 +311,7 @@ MODULE mod_calendar
          REAL(DP), INTENT(IN)  :: tv
 
          ! Now update the time and date
-         dateSec   = tv
+         dateSec   = INT(tv)
          dateMin   = startMin
          dateHour  = startHour
          dateDay   = startDay
